A few changes so that clang doesn't throw a shitload of warnings. For shit and giggles I also enabled -Werror in the main Makefile.

+++ drivers/vhost/vhost.c
@@ -1437,7 +1437,7 @@ long vhost_vring_ioctl(struct vhost_dev *d, int ioctl, void __user *argp)
 		/* Forget the cached index value. */
 		vq->avail_idx = vq->last_avail_idx;
 		break;
-	case VHOST_GET_VRING_BASE:
+	case (int)VHOST_GET_VRING_BASE:
 		s.index = idx;
 		s.num = vq->last_avail_idx;
 		if (copy_to_user(argp, &s, sizeof s))
